CLE_004: New game hangman
